Create a bootable disk image atop the existing freestanding binary.
We also simplify the build process by adding `bootloader` as a
dependency in order to compile the minimal kernel, compile the
bootloader, then link the two post-compilation in order to produce a
bootable disk image--all with a simple `$ cargo run`.

README updates reflect the new simplified build steps, as well as
include a screenshot of the running QEMU machine. Yay!